server/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
pymongo==4.6.2
python-dotenv==1.0.1
